Refactor: Split result module into types, errors, and functions
