+++ src/result/index.ts
@@ -1,0 +1,4 @@
+export * from './types';
+export * from './errors';
+export * from './functions';
+
